[chore] Added plantUML

+++ lib/04_Inheritance/rectangle/point.dart
@@ -1,6 +1,6 @@
 class Point {
-  int x;
-  int y;
+  double x;
+  double y;
 
   Point(this.x, this.y);
 }
